// src/app/models/usuario.ts
export interface DatosUsuario {
  cPerCodigo: string;
  cPerApellido: string;
  cPerNombre: string;
  cPerUsuCodigo: string;
  cPerEmail: string;
  cPerTipoDoc: string;
  cPerNroDoc: string;
  nTipo: number;
  cTipoDesc: string;
  cCargo: string;
  nRol: number;
  cToken: string;
  bLegajo: boolean;
}